Move getUncategorizedSenders into its own file
